restructure and merge tasks and sorter
