+++ src/provider/tasks.ts
@@ -1,9 +1,11 @@
-import {IStatistics, ICategoricalStatistics, IDateStatistics, IAdvancedBoxPlotData, dateStatsBuilder, normalizedStatsBuilder, categoricalStatsBuilder, boxplotBuilder, getNumberOfBins} from '../internal/math';
-import {IDataRow, INumberColumn, IDateColumn, ISetColumn, IOrderedGroup, IndicesArray, Ranking, ICategoricalLikeColumn, IGroup, NumberColumn, OrdinalColumn, ImpositionCompositeColumn, CategoricalColumn, DateColumn, isCategoricalLikeColumn, isNumberColumn, isDateColumn} from '../model';
+import {abortAble, abortAbleAll, IAbortAblePromise} from 'lineupengine';
+import {IForEachAble, lazySeq} from '../internal/interable';
+import {boxplotBuilder, categoricalStatsBuilder, dateStatsBuilder, IAdvancedBoxPlotData, ICategoricalStatistics, IDateStatistics, IStatistics, normalizedStatsBuilder} from '../internal/math';
+import {ANOTHER_ROUND} from '../internal/scheduler';
+import {CategoricalColumn, DateColumn, ICategoricalLikeColumn, IDataRow, IDateColumn, IGroup, ImpositionCompositeColumn, IndicesArray, INumberColumn, NumberColumn, OrdinalColumn, Ranking} from '../model';
 import Column, {ICompareValue} from '../model/Column';
-import {ISequence, lazySeq, IForEachAble} from '../internal/interable';
-import {IAbortAblePromise, ABORTED, abortAbleAll, abortAble} from 'lineupengine';
-import TaskScheduler, {ANOTHER_ROUND, oneShotIterator} from '../internal/scheduler';
+import {IRenderTask, IRenderTasks} from '../renderer/interfaces';
+import {CompareLookup} from './sort';
 
 export {IAbortAblePromise} from 'lineupengine';
 
@@ -16,11 +18,10 @@ export {IAbortAblePromise} from 'lineupengine';
 // or transfer to the worker threads to compute the hists async and keep in sync
 // also use the value cache for computing the comparevalue and groupcompareavalue
 
-export interface IRenderTask<T> {
-  then<U = void>(onfullfilled: (value: T | symbol) => U): U | IAbortAblePromise<U>;
-}
-
-class TaskNow<T> implements IRenderTask<T> {
+/**
+ * @internal
+ */
+export class TaskNow<T> implements IRenderTask<T> {
   constructor(public readonly v: T) {
 
   }
@@ -30,11 +31,17 @@ class TaskNow<T> implements IRenderTask<T> {
   }
 }
 
-function taskNow<T>(v: T) {
+/**
+ * @internal
+ */
+export function taskNow<T>(v: T) {
   return new TaskNow(v);
 }
 
-class TaskLater<T> implements IRenderTask<T> {
+/**
+ * @internal
+ */
+export class TaskLater<T> implements IRenderTask<T> {
   constructor(public readonly v: IAbortAblePromise<T>) {
 
   }
@@ -45,7 +52,10 @@ class TaskLater<T> implements IRenderTask<T> {
   }
 }
 
-function taskLater<T>(v: IAbortAblePromise<T>) {
+/**
+ * @internal
+ */
+export function taskLater<T>(v: IAbortAblePromise<T>) {
   return new TaskLater(v);
 }
 
@@ -56,20 +66,7 @@ export function tasksAll<T>(tasks: IRenderTask<T>[]): IRenderTask<T[]> {
   return taskLater(abortAbleAll((<(TaskNow<T> | TaskLater<T>)[]>tasks).map((d) => d.v)));
 }
 
-export interface IRenderTasks {
-  groupRows<T>(col: Column, group: IOrderedGroup, key: string, compute: (rows: ISequence<IDataRow>) => T): IRenderTask<T>;
-  groupExampleRows<T>(col: Column, group: IOrderedGroup, key: string, compute: (rows: ISequence<IDataRow>) => T): IRenderTask<T>;
 
-  groupBoxPlotStats(col: Column & INumberColumn, group: IOrderedGroup, raw?: boolean): IRenderTask<{group: IAdvancedBoxPlotData, summary: IAdvancedBoxPlotData, data: IAdvancedBoxPlotData}>;
-  groupNumberStats(col: Column & INumberColumn, group: IOrderedGroup, raw?: boolean): IRenderTask<{group: IStatistics, summary: IStatistics, data: IStatistics}>;
-  groupCategoricalStats(col: Column & ICategoricalLikeColumn, group: IOrderedGroup): IRenderTask<{group: ICategoricalStatistics, summary: ICategoricalStatistics, data: ICategoricalStatistics}>;
-  groupDateStats(col: Column & IDateColumn, group: IOrderedGroup): IRenderTask<{group: IDateStatistics, summary: IDateStatistics, data: IDateStatistics}>;
-
-  summaryBoxPlotStats(col: Column & INumberColumn, raw?: boolean): IRenderTask<{summary: IAdvancedBoxPlotData, data: IAdvancedBoxPlotData}>;
-  summaryNumberStats(col: Column & INumberColumn, raw?: boolean): IRenderTask<{summary: IStatistics, data: IStatistics}>;
-  summaryCategoricalStats(col: Column & ICategoricalLikeColumn): IRenderTask<{summary: ICategoricalStatistics, data: ICategoricalStatistics}>;
-  summaryDateStats(col: Column & IDateColumn): IRenderTask<{summary: IDateStatistics, data: IDateStatistics}>;
-}
 
 export interface IRenderTaskExectutor extends IRenderTasks {
   setData(data: IDataRow[]): void;
@@ -83,14 +80,23 @@ export interface IRenderTaskExectutor extends IRenderTasks {
   preComputeData(ranking: Ranking): void;
   copyData2Summary(ranking: Ranking): void;
   copyCache(col: Column, from: Column): void;
+
+  sort(indices: IndicesArray, singleCall: boolean, lookups?: CompareLookup): Promise<IndicesArray>;
+  terminate(): void;
 }
 
-class MultiIndices {
+/**
+ * @internal
+ */
+export class MultiIndices {
   constructor(public readonly indices: IndicesArray[]) {
 
   }
 }
 
+/**
+ * @internal
+ */
 export class ARenderTasks {
   protected readonly byIndex = (i: number) => this.data[i];
 
@@ -192,518 +198,4 @@ export class ARenderTasks {
     return this.builderForEach(b, order, (row: IDataRow) => col.iterCategory(row), build);
   }
 
-}
-
-export class DirectRenderTasks extends ARenderTasks implements IRenderTaskExectutor {
-
-  protected readonly cache = new Map<string, any>();
-
-  setData(data: IDataRow[]) {
-    this.data = data;
-    this.cache.clear();
-  }
-
-
-  dirtyColumn(col: Column, type: 'data' | 'summary' | 'group') {
-    if (type === 'group') {
-      return; // not cached
-    }
-    this.cache.delete(`${col.id}:summary`);
-    this.cache.delete(`${col.id}:summary:raw`);
-    this.cache.delete(`${col.id}:summary:b`);
-    this.cache.delete(`${col.id}:summary:braw`);
-
-    if (type === 'summary') {
-      return;
-    }
-    this.cache.delete(`${col.id}:data`);
-    this.cache.delete(`${col.id}:data:raw`);
-    this.cache.delete(`${col.id}:data:b`);
-    this.cache.delete(`${col.id}:data:braw`);
-  }
-
-  dirtyRanking(ranking: Ranking, type: 'data' | 'summary' | 'group') {
-    for (const col of ranking.flatColumns) {
-      this.dirtyColumn(col, type);
-    }
-  }
-
-  preCompute() {
-    // dummy
-  }
-
-  preComputeData() {
-    // dummy
-  }
-
-  preComputeCol() {
-    // dummy
-  }
-
-  copyData2Summary() {
-    // dummy
-  }
-
-  copyCache(col: Column, from: Column) {
-    const fromPrefix = `${from.id}:`;
-
-    for (const key of Array.from(this.cache.keys()).sort()) {
-      if (!key.startsWith(fromPrefix)) {
-        continue;
-      }
-      const tkey = `${col.id}:${key.slice(fromPrefix.length)}`;
-      this.cache.set(tkey, this.cache.get(key)!);
-    }
-  }
-
-  groupCompare(ranking: Ranking, group: IGroup, rows: IndicesArray) {
-    return taskNow(ranking.toGroupCompareValue(this.byOrder(rows), group));
-  }
-
-  groupRows<T>(_col: Column, group: IOrderedGroup, _key: string, compute: (rows: ISequence<IDataRow>) => T) {
-    return taskNow(compute(this.byOrder(group.order)));
-  }
-
-  groupExampleRows<T>(_col: Column, group: IOrderedGroup, _key: string, compute: (rows: ISequence<IDataRow>) => T) {
-    return taskNow(compute(this.byOrder(group.order.slice(0, 5))));
-  }
-
-  groupBoxPlotStats(col: Column & INumberColumn, group: IOrderedGroup, raw?: boolean) {
-    const {summary, data} = this.summaryBoxPlotStatsD(col, raw);
-    return taskNow({group: this.boxplotBuilder(group.order, col, raw).next(Infinity).value!, summary, data});
-  }
-
-  groupNumberStats(col: Column & INumberColumn, group: IOrderedGroup, raw?: boolean) {
-    const {summary, data} = this.summaryNumberStatsD(col, raw);
-    return taskNow({group: this.normalizedStatsBuilder(group.order, col, summary.hist.length, raw).next(Infinity).value!, summary, data});
-  }
-
-  groupCategoricalStats(col: Column & ISetColumn, group: IOrderedGroup) {
-    const {summary, data} = this.summaryCategoricalStatsD(col);
-    return taskNow({group: this.categoricalStatsBuilder(group.order, col).next(Infinity).value!, summary, data});
-  }
-
-  groupDateStats(col: Column & IDateColumn, group: IOrderedGroup) {
-    const {summary, data} = this.summaryDateStatsD(col);
-    return taskNow({group: this.dateStatsBuilder(group.order, col, summary).next(Infinity).value!, summary, data});
-  }
-
-  summaryBoxPlotStats(col: Column & INumberColumn, raw?: boolean) {
-    return taskNow(this.summaryBoxPlotStatsD(col, raw));
-  }
-
-  summaryNumberStats(col: Column & INumberColumn, raw?: boolean) {
-    return taskNow(this.summaryNumberStatsD(col, raw));
-  }
-
-  summaryCategoricalStats(col: Column & ISetColumn) {
-    return taskNow(this.summaryCategoricalStatsD(col));
-  }
-
-  summaryDateStats(col: Column & IDateColumn) {
-    return taskNow(this.summaryDateStatsD(col));
-  }
-
-  private summaryNumberStatsD(col: Column & INumberColumn, raw?: boolean) {
-    return this.cached('summary', col, () => {
-      const ranking = col.findMyRanker()!.getOrder();
-      const data = this.dataNumberStats(col, raw);
-      return {summary: this.normalizedStatsBuilder(ranking, col, data.hist.length, raw).next(Infinity).value!, data};
-    }, raw ? ':raw' : '', col.findMyRanker()!.getOrderLength() === 0);
-  }
-
-  private summaryBoxPlotStatsD(col: Column & INumberColumn, raw?: boolean) {
-    return this.cached('summary', col, () => {
-      const ranking = col.findMyRanker()!.getOrder();
-      const data = this.dataBoxPlotStats(col, raw);
-      return {summary: this.boxplotBuilder(ranking, col, raw).next(Infinity).value!, data};
-    }, raw ? ':braw' : ':b' , col.findMyRanker()!.getOrderLength() === 0);
-  }
-
-  private summaryCategoricalStatsD(col: Column & ISetColumn) {
-    return this.cached('summary', col, () => {
-      const ranking = col.findMyRanker()!.getOrder();
-      const data = this.dataCategoricalStats(col);
-      return {summary: this.categoricalStatsBuilder(ranking, col).next(Infinity).value!, data};
-    }, '', col.findMyRanker()!.getOrderLength() === 0);
-  }
-
-  private summaryDateStatsD(col: Column & IDateColumn) {
-    return this.cached('summary', col, () => {
-      const ranking = col.findMyRanker()!.getOrder();
-      const data = this.dataDateStats(col);
-      return {summary: this.dateStatsBuilder(ranking, col, data).next(Infinity).value!, data};
-    }, '', col.findMyRanker()!.getOrderLength() === 0);
-  }
-
-  private cached<T>(prefix: string, col: Column, creator: () => T, suffix: string = '', dontCache = false): T {
-    const key = `${col.id}:${prefix}${suffix}`;
-    if (this.cache.has(key)) {
-      return this.cache.get(key)!;
-    }
-    const s = creator();
-    if (!dontCache) {
-      this.cache.set(key, s);
-    }
-    return s;
-  }
-
-  dataBoxPlotStats(col: Column & INumberColumn, raw?: boolean) {
-    return this.cached('data', col, () => this.boxplotBuilder(null, col, raw).next(Infinity).value!, raw ? ':braw' : ':b');
-  }
-
-  dataNumberStats(col: Column & INumberColumn, raw?: boolean) {
-    return this.cached('data', col, () => this.normalizedStatsBuilder(null, col, getNumberOfBins(this.data.length), raw).next(Infinity).value!, raw ? ':raw' : '');
-  }
-
-  dataCategoricalStats(col: Column & ISetColumn) {
-    return this.cached('data', col, () => this.categoricalStatsBuilder(null, col).next(Infinity).value!);
-  }
-
-  dataDateStats(col: Column & IDateColumn) {
-    return this.cached('data', col, () => this.dateStatsBuilder(null, col).next(Infinity).value!);
-  }
-
-}
-
-
-export class ScheduleRenderTasks extends ARenderTasks implements IRenderTaskExectutor {
-
-  private readonly cache = new Map<string, IRenderTask<any>>();
-  private readonly tasks = new TaskScheduler();
-
-  setData(data: IDataRow[]) {
-    this.data = data;
-    this.cache.clear();
-    this.tasks.clear();
-  }
-
-  dirtyColumn(col: Column, type: 'data' | 'group' | 'summary') {
-    // order designed such that first groups, then summaries, then data is deleted
-
-    for (const key of Array.from(this.cache.keys()).sort()) {
-      // data = all
-      // summary = summary + group
-      // group = group only
-      if ((type === 'data' && key.startsWith(`${col.id}:`) ||
-        (type === 'summary' && key.startsWith(`${col.id}:b:summary:`)) ||
-        (key.startsWith(`${col.id}:a:group`)))) {
-        this.cache.delete(key);
-        this.tasks.abort(key);
-      }
-    }
-  }
-
-  dirtyRanking(ranking: Ranking, type: 'data' | 'group' | 'summary') {
-    const cols = ranking.flatColumns;
-
-    let checker: ((key: string) => boolean)[];
-    switch (type) {
-      case 'group':
-        checker = cols.map((col) => (key: string) => key.startsWith(`${col.id}:a:group`));
-        break;
-      case 'summary':
-        checker = cols.map((col) => (key: string) => key.startsWith(`${col.id}:b:summary`) || key.startsWith(`${col.id}:a:group`));
-        break;
-      case 'data':
-      default:
-        checker = cols.map((col) => (key: string) => key.startsWith(`${col.id}:`));
-        break;
-    }
-    for (const key of Array.from(this.cache.keys()).sort()) {
-      if (checker.some((f) => f(key))) {
-        this.cache.delete(key);
-        this.tasks.abort(key);
-      }
-    }
-    // group compare tasks
-    this.tasks.abortAll((t) => t.id.startsWith(`r${ranking.id}:`));
-  }
-
-  preCompute(ranking: Ranking, groups: {rows: IndicesArray, group: IGroup}[]) {
-    if (groups.length === 0) {
-      return;
-    }
-    const cols = ranking.flatColumns;
-    if (groups.length === 1) {
-      const {group, rows} = groups[0];
-      const multi = new MultiIndices([rows]);
-      for (const col of cols) {
-        if (isCategoricalLikeColumn(col)) {
-          this.summaryCategoricalStats(col, multi);
-        } else if (isNumberColumn(col)) {
-          this.summaryNumberStats(col, false, multi);
-        } else if (isDateColumn(col)) {
-          this.summaryDateStats(col, multi);
-        } else {
-          continue;
-        }
-        // copy from summary to group and create proper structure
-        this.chainCopy(`${col.id}:a:group:${group.name}`, this.cache.get(`${col.id}:b:summary`)!, (v: {summary: any, data: any}) => ({group: v.summary, summary: v.summary, data: v.data}));
-      }
-      return;
-    }
-
-    const ogroups = groups.map(({rows, group}) => Object.assign({order: rows}, group));
-    const full = new MultiIndices(groups.map((d) => d.rows));
-    for (const col of cols) {
-      if (isCategoricalLikeColumn(col)) {
-        this.summaryCategoricalStats(col, full);
-        for (const g of ogroups) {
-          this.groupCategoricalStats(col, g);
-        }
-      } else if (isNumberColumn(col)) {
-        this.summaryNumberStats(col, false, full);
-        for (const g of ogroups) {
-          this.groupNumberStats(col, g, false);
-        }
-      } else if (isDateColumn(col)) {
-        this.summaryDateStats(col, full);
-        for (const g of ogroups) {
-          this.groupDateStats(col, g);
-        }
-      }
-    }
-  }
-
-  preComputeData(ranking: Ranking) {
-    for (const col of ranking.flatColumns) {
-      if (isCategoricalLikeColumn(col)) {
-        this.dataCategoricalStats(col);
-      } else if (isNumberColumn(col)) {
-        this.dataNumberStats(col);
-      } else if (isDateColumn(col)) {
-        this.dataDateStats(col);
-      }
-    }
-  }
-
-  preComputeCol(col: Column) {
-    const ranking = col.findMyRanker();
-
-    if (isCategoricalLikeColumn(col)) {
-      this.dataCategoricalStats(col);
-      if (!ranking) {
-        return;
-      }
-      this.summaryCategoricalStats(col);
-      for (const group of ranking.getGroups()) {
-        this.groupCategoricalStats(col, group);
-      }
-      return;
-    }
-
-    if (isNumberColumn(col)) {
-      this.dataNumberStats(col);
-
-      if (!ranking) {
-        return;
-      }
-      this.summaryNumberStats(col);
-      for (const group of ranking.getGroups()) {
-        this.groupNumberStats(col, group);
-      }
-      return;
-    }
-
-    if (!isDateColumn(col)) {
-      return;
-    }
-
-    this.dataDateStats(col);
-
-    if (!ranking) {
-      return;
-    }
-    this.summaryDateStats(col);
-    for (const group of ranking.getGroups()) {
-      this.groupDateStats(col, group);
-    }
-  }
-
-  copyData2Summary(ranking: Ranking) {
-    for (const col of ranking.flatColumns) {
-      if (isCategoricalLikeColumn(col)) {
-        this.dataCategoricalStats(col);
-      } else if (isNumberColumn(col)) {
-        this.dataNumberStats(col);
-      } else if (isDateColumn(col)) {
-        this.dataDateStats(col);
-      } else {
-        continue;
-      }
-      // copy from data to summary and create proper structure
-      this.chainCopy(`${col.id}:b:summary`, this.cache.get(`${col.id}:c:data`)!, (data: any) => ({summary: data, data}));
-    }
-  }
-
-  copyCache(col: Column, from: Column) {
-    const fromPrefix = `${from.id}:`;
-
-    for (const key of Array.from(this.cache.keys()).sort()) {
-      if (!key.startsWith(fromPrefix)) {
-        continue;
-      }
-      const tkey = `${col.id}:${key.slice(fromPrefix.length)}`;
-      this.chainCopy(tkey, this.cache.get(key)!, (data: any) => data);
-    }
-  }
-
-  groupCompare(ranking: Ranking, group: IGroup, rows: IndicesArray) {
-    return taskLater(this.tasks.push(`r${ranking.id}:${group.name}`, () => ranking.toGroupCompareValue(this.byOrder(rows), group)));
-  }
-
-  groupRows<T>(col: Column, group: IOrderedGroup, key: string, compute: (rows: ISequence<IDataRow>) => T) {
-    return this.cached(`${col.id}:a:group:${group.name}:${key}`, group.order.length === 0, oneShotIterator(() => compute(this.byOrder(group.order))));
-  }
-
-  groupExampleRows<T>(_col: Column, group: IOrderedGroup, _key: string, compute: (rows: ISequence<IDataRow>) => T) {
-    return taskNow(compute(this.byOrder(group.order.slice(0, 5))));
-  }
-
-  groupBoxPlotStats(col: Column & INumberColumn, group: IOrderedGroup, raw?: boolean) {
-    return this.chain(`${col.id}:a:group:${group.name}${raw ? ':braw' : ':b'}`, this.summaryBoxPlotStats(col, raw), ({summary, data}) => {
-      return this.boxplotBuilder(group.order, col, raw, (group) => ({group, summary, data}));
-    });
-  }
-
-  groupNumberStats(col: Column & INumberColumn, group: IOrderedGroup, raw?: boolean) {
-    return this.chain(`${col.id}:a:group:${group.name}${raw ? ':raw' : ''}`, this.summaryNumberStats(col, raw), ({summary, data}) => {
-      return this.normalizedStatsBuilder(group.order, col, summary.hist.length, raw, (group) => ({group, summary, data}));
-    });
-  }
-
-  groupCategoricalStats(col: Column & ICategoricalLikeColumn, group: IOrderedGroup) {
-    return this.chain(`${col.id}:a:group:${group.name}`, this.summaryCategoricalStats(col), ({summary, data}) => {
-      return this.categoricalStatsBuilder(group.order, col, (group) => ({group, summary, data}));
-    });
-  }
-
-  groupDateStats(col: Column & IDateColumn, group: IOrderedGroup) {
-    return this.chain(`${col.id}:a:group:${group.name}`, this.summaryDateStats(col), ({summary, data}) => {
-      return this.dateStatsBuilder(group.order, col, summary, (group) => ({group, summary, data}));
-    });
-  }
-
-  summaryBoxPlotStats(col: Column & INumberColumn, raw?: boolean, order?: MultiIndices) {
-    return this.chain(`${col.id}:b:summary${raw ? ':braw' : ':b'}`, this.dataBoxPlotStats(col, raw), (data) => {
-      return this.boxplotBuilder(order ? order : col.findMyRanker()!.getOrder(), col, raw, (summary) => ({summary, data}));
-    });
-  }
-
-  summaryNumberStats(col: Column & INumberColumn, raw?: boolean, order?: MultiIndices) {
-    return this.chain(`${col.id}:b:summary${raw ? ':raw' : ''}`, this.dataNumberStats(col, raw), (data) => {
-      return this.normalizedStatsBuilder(order ? order : col.findMyRanker()!.getOrder(), col, data.hist.length, raw, (summary) => ({summary, data}));
-    });
-  }
-
-  summaryCategoricalStats(col: Column & ICategoricalLikeColumn, order?: MultiIndices) {
-    return this.chain(`${col.id}:b:summary`, this.dataCategoricalStats(col), (data) => {
-      return this.categoricalStatsBuilder(order ? order : col.findMyRanker()!.getOrder(), col, (summary) => ({summary, data}));
-    });
-  }
-
-  summaryDateStats(col: Column & IDateColumn, order?: MultiIndices) {
-    return this.chain(`${col.id}:b:summary`, this.dataDateStats(col), (data) => {
-      return this.dateStatsBuilder(order ? order : col.findMyRanker()!.getOrder(), col, data, (summary) => ({summary, data}));
-    });
-  }
-
-  private cached<T>(key: string, dontCache: boolean, it: Iterator<T | null>): IRenderTask<T> {
-    if (this.cache.has(key) && !dontCache) {
-      return this.cache.get(key)!;
-    }
-
-    const task = this.tasks.pushMulti(key, it);
-    const s = taskLater(task);
-    if (!dontCache) {
-      this.cache.set(key, s);
-    }
-    task.then((r) => {
-      if (typeof r === 'symbol') {
-        return;
-      }
-      if (this.cache.get(key) === s) {
-        // still same value replace with faster version
-        this.cache.set(key, taskNow(r));
-      }
-    });
-    return s;
-  }
-
-  private chain<T, U>(key: string, task: IRenderTask<T>, creator: (data: T) => Iterator<U | null>): IRenderTask<U> {
-    if (this.cache.has(key)) {
-      return this.cache.get(key)!;
-    }
-    if (task instanceof TaskNow) {
-      return this.cached(key, false, creator(task.v));
-    }
-
-    const v = (<TaskLater<T>>task).v;
-    const subTask = v.then((data) => {
-      if (typeof data === 'symbol') {
-        return ABORTED;
-      }
-      return this.tasks.pushMulti(key, creator(data));
-    });
-    const s = taskLater(subTask);
-    this.cache.set(key, s);
-    subTask.then((r) => {
-      if (typeof r === 'symbol') {
-        return;
-      }
-      if (this.cache.get(key) === s) {
-        // still same value replace with faster version
-        this.cache.set(key, taskNow(r));
-      }
-    });
-    return s;
-  }
-
-  private chainCopy<T, U>(key: string, task: IRenderTask<T>, creator: (data: T) => U): IRenderTask<U> {
-    if (this.cache.has(key)) {
-      return this.cache.get(key)!;
-    }
-    if (task instanceof TaskNow) {
-      const subTask = taskNow(creator(task.v));
-      this.cache.set(key, subTask);
-      return subTask;
-    }
-
-    const v = (<TaskLater<T>>task).v;
-    const subTask = v.then((data) => {
-      if (typeof data === 'symbol') {
-        return ABORTED;
-      }
-      return creator(data);
-    });
-    const s = taskLater(subTask);
-    this.cache.set(key, s);
-    subTask.then((r) => {
-      if (typeof r === 'symbol') {
-        return;
-      }
-      if (this.cache.get(key) === s) {
-        // still same value replace with faster version
-        this.cache.set(key, taskNow(r));
-      }
-    });
-    return s;
-  }
-
-  dataBoxPlotStats(col: Column & INumberColumn, raw?: boolean) {
-    return this.cached(`${col.id}:c:data${raw ? ':braw' : ':b'}`, this.data.length === 0, this.boxplotBuilder<IAdvancedBoxPlotData>(null, col, raw));
-  }
-
-  dataNumberStats(col: Column & INumberColumn, raw?: boolean) {
-    return this.cached(`${col.id}:c:data${raw ? ':raw' : ''}`, this.data.length === 0, this.normalizedStatsBuilder<IStatistics>(null, col, getNumberOfBins(this.data.length), raw));
-  }
-
-  dataCategoricalStats(col: Column & ICategoricalLikeColumn) {
-    return this.cached(`${col.id}:c:data`, this.data.length === 0, this.categoricalStatsBuilder<ICategoricalStatistics>(null, col));
-  }
-
-  dataDateStats(col: Column & IDateColumn) {
-    return this.cached(`${col.id}:c:data`, this.data.length === 0, this.dateStatsBuilder<IDateStatistics>(null, col));
-  }
 }
